refactor: helicityformalism -> helicity_formalism
