+++ tests/data/test_generate.py
@@ -1,7 +1,7 @@
 import pytest
 
 from tensorwaves.data.generate import generate_phsp
-from tensorwaves.physics.helicityformalism.kinematics import (
+from tensorwaves.physics.helicity_formalism.kinematics import (
     HelicityKinematics,
     ParticleReactionKinematicsInfo,
 )
